Create README.md
In this project, we'll be working with data from the S&P500 Index. 
The S&P500 is a stock market index.

We'll be using historical data on the price of the S&P500 Index 
to make predictions about future prices. Predicting whether an index will go 
up or down will help us forecast how the stock market as a whole will perform.

We'll be working with a csv file containing index prices. 
Each row in the file contains a daily record of the price of the S&P500 Index 
from 1950 to 2015. The dataset is stored in sphist.csv'''